pyproject.toml
[tool.poetry]
name = "python-pype"
version = "0.0.63"
description = "Command-line pipes in Python"
authors = ["author <author@example.com>"]
readme = "README.rst"
packages = [
         {include = "pype"}
]

[tool.poetry.dependencies]
python = "^3.6"

click = "*"
toolz = "*"
attrs = "*"
click-default-group = "*"
parso = "*"
bump2version = "*"
trio = "*"
async_generator = "*"
async_exit_stack = "*"
appdirs = "*"
toml = "*"

[tool.poetry.dev-dependencies]

bump2version = "*"
coverage = "*"
hypothesis = "*"
pylint = "*"
pytest = "*"
requests = "*"
aiohttp = "*"
jedi = "*"
"autopep8" = "*"
yapf = "*"
parso = "*"
twine = "*"
vulture = "*"
treq = "*"
poetry = "*"
twisted = "*"
asks = "*"


[tool.poetry.scripts]

pype = 'pype:cli.cli'


[tool.poetry.plugins.pype_plugins]
"basic" = "pype.plugins"
